+++ sushi_lang/backend/statements/utils.py
@@ -36,7 +36,7 @@ def emit_condition(codegen: 'LLVMCodegen', expr) -> 'ir.Value':
     value = codegen.expressions.emit_expr(expr)
     cond = codegen.utils.as_i1(value)
 
-    if expression_is_temporary(expr):
+    if expression_is_temporary(codegen, expr):
         enum_type = (infer_generic_enum_type(codegen, expr, value, "Result<")
                      or infer_generic_enum_type(codegen, expr, value, "Maybe<"))
         if enum_type is not None: